rename base_algo to agent
